+++ static/__tests__/iframe/DataViewer-base-test.jsx
@@ -143,6 +143,7 @@ describe("DataViewer iframe tests", () => {
           "Rename",
           "Replacements",
           "Type Conversion",
+          "Duplicates",
           "Describe",
           "Column Analysis",
           "Formats",